Fix #863: move ProcessRunner to wanaku-capabilities-java-sdk
