+++ cli/pom.xml
@@ -74,6 +74,10 @@
             <artifactId>core-util</artifactId>
             <version>${project.version}</version>
         </dependency>
+        <dependency>
+            <groupId>ai.wanaku.sdk</groupId>
+            <artifactId>capabilities-common</artifactId>
+        </dependency>
 
         <dependency>
             <groupId>ai.wanaku</groupId>
