+++ cli/src/main/java/ai/wanaku/cli/main/commands/capabilities/CapabilitiesBase.java
@@ -3,8 +3,8 @@ package ai.wanaku.cli.main.commands.capabilities;
 import java.io.File;
 import org.jboss.logging.Logger;
 import ai.wanaku.capabilities.sdk.api.exceptions.WanakuException;
+import ai.wanaku.capabilities.sdk.common.ProcessRunner;
 import ai.wanaku.cli.main.commands.BaseCommand;
-import ai.wanaku.core.util.ProcessRunner;
 import ai.wanaku.core.util.VersionHelper;
 import picocli.CommandLine;
