+++ core/core-util/pom.xml
@@ -17,6 +17,10 @@
             <artifactId>capabilities-api</artifactId>
         </dependency>
         <dependency>
+            <groupId>ai.wanaku.sdk</groupId>
+            <artifactId>capabilities-common</artifactId>
+        </dependency>
+        <dependency>
             <groupId>com.fasterxml.jackson.core</groupId>
             <artifactId>jackson-databind</artifactId>
             <version>${jackson.version}</version>
